+++ goondex/makefile
@@ -2,4 +2,4 @@ run:
 	@go run ./cmd/cli/main.go
 
 test:
-	@go test goondex/engine goondex/engine/storage goondex/web goondex/index goondex/crawler
+	@go test goondex/engine goondex/webpages/storage goondex/webpages goondex/index goondex/crawler goondex/bitree
