--- goondex/pkg/crawler/interface.go
@@ -1,7 +1,0 @@
-package crawler
-
-import "goondex/web"
-
-type Interface interface {
-	Scan(string, int) ([]web.Page, error)
-}
