+++ goondex/pkg/crawler/site/scan.go
