+++ goondex/pkg/index/index.go
@@ -1,15 +1,15 @@
 package index
 
 import (
-	"goondex/web"
+	"goondex/webpages"
 	"strings"
 )
 
 type Index struct {
-	index map[string][]int
+	index map[string][]uint32
 }
 
-func (ind *Index) Add(page web.Page) {
+func (ind *Index) Add(page webpages.Page) {
 	tokens := extractTokens(page.Title)
 	for _, token := range tokens {
 		lowerToken := strings.ToLower(token)
@@ -17,13 +17,13 @@ func (ind *Index) Add(page web.Page) {
 	}
 }
 
-func (ind *Index) Search(query string) []int {
+func (ind *Index) Search(query string) []uint32 {
 	return ind.index[strings.ToLower(query)]
 }
 
 func New() *Index {
 	return &Index{
-		index: make(map[string][]int),
+		index: make(map[string][]uint32),
 	}
 }
