--- goondex/pkg/web/go.mod
@@ -1,3 +1,0 @@
-module web
-
-go 1.15
