--- goondex/pkg/web/page.go
@@ -1,8 +1,0 @@
-package web
-
-// Описание страницы
-type Page struct {
-	URL   string
-	Title string
-	ID    int
-}
