--- review/goondex/makefile
@@ -1,5 +1,0 @@
-run:
-	@go run ./cmd/cli/main.go
-
-test:
-	@go test goondex/engine goondex/webpages/storage goondex/webpages goondex/index goondex/crawler goondex/bitree
