--- review/goondex/pkg/index/go.mod
@@ -1,3 +1,0 @@
-module index
-
-go 1.15
