--- review/goondex/pkg/index/interface.go
@@ -1,9 +1,0 @@
-package index
-
-import "goondex/webpages"
-
-// Позволяет добавлять документы в индекс и искать по индексу
-type Interface interface {
-	Add(page webpages.Page)
-	Search(query string) []uint32
-}
