--- review/goondex/pkg/webpages/storage/go.mod
@@ -1,3 +1,0 @@
-module storage
-
-go 1.15
